requirements.txt
# python requirements
numpy>=1.19.1
opencv-python>=4.4.0.42
Pillow>=7.2.0
scipy>=1.5.2
scikit-learn>=0.23.2
dlib>=19.21.0
pygame>=1.9.6
PyOpenGL>=3.1.5
trimesh>=3.8.8
